Cache the http responses in QML's SQLite LocalStorage.
I constantly restart plasmashell, so my widget listed 10 repos, I'd
run into my rate limit after restarting plasmashell 7 times in a
single hour.

This code should be updated to handle network failure (eg: offline),
in which case we should still use the cached result even if it has
expired. This code would be less useful for this widget (since it's
just links to a website), but should be useful for eventcalendar.